Separate password validation from other sign up form validations
