+++ lib/validation/models/password_validation.dart
@@ -1,0 +1,11 @@
+class ValidatedPassword {
+  final String? value;
+
+  ValidatedPassword(this.value);
+}
+
+class PasswordValidationError {
+  final String? error;
+
+  PasswordValidationError(this.error);
+}
