+++ lib/validation/models/validation_item.dart
@@ -4,25 +4,3 @@ class ValidationItem {
 
   ValidationItem(this.value, this.error);
 }
-
-class ValidatedPassword {
-  final String? value;
-
-  ValidatedPassword(this.value);
-}
-
-class PasswordValidationError {
-  final String? error;
-
-  PasswordValidationError(this.error);
-}
-
-// void confirmPasswordField(String pwd) {
-//   if (pwd.length < 8
-//       && !RegExp(r'(?=.*[A-Z])').hasMatch(pwd)
-//       && !RegExp(r'(?=.*[a-z])').hasMatch(pwd)
-//       && !RegExp(r'(?=.*\d)').hasMatch(pwd)
-//   ) {
-//     _confirmPassword = ValidationItem(pwd, null);
-//   }
-// }
